# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Infinite Duo - Couples Game Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  name: string;
  avatar: string;
  totalPoints: number;
  weeklyPoints: number;
  gamesPlayed: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePlayerBody {
  name: string;
  avatar: string;
}

export interface LeaderboardEntry {
  playerId: number;
  playerName: string;
  avatar: string;
  weeklyPoints: number;
  totalPoints: number;
  rank: number;
}

export interface LeaderboardResponse {
  weeklyLeader?: LeaderboardEntry;
  entries: LeaderboardEntry[];
  weekEndsAt: string;
}

export interface RelationshipLevel {
  level: number;
  levelName: string;
  currentXp: number;
  xpToNextLevel: number;
  totalXp: number;
  progressPercent: number;
}

export interface GameStats {
  totalGamesPlayed: number;
  predictionsWon: number;
  predictionsTotal: number;
  shadowDuelTopScore: number;
  syncUpSessionsPlayed: number;
  currentStreak: number;
}

export type PredictionStatus =
  (typeof PredictionStatus)[keyof typeof PredictionStatus];

export const PredictionStatus = {
  pending: "pending",
  correct: "correct",
  wrong: "wrong",
} as const;

export interface Prediction {
  id: number;
  playerId: number;
  playerName: string;
  question: string;
  prediction: string;
  betPoints: number;
  status: PredictionStatus;
  /** @nullable */
  resolvedAt?: string | null;
  createdAt: string;
}

export interface CreatePredictionBody {
  playerId: number;
  question: string;
  prediction: string;
  betPoints: number;
}

export interface ResolvePredictionBody {
  correct: boolean;
}

export interface PredictionHistory {
  totalPredictions: number;
  correctPredictions: number;
  winRate: number;
  recentPredictions: Prediction[];
}

export interface ShadowDuelScore {
  id: number;
  playerId: number;
  playerName: string;
  avatar: string;
  score: number;
  level: number;
  combo: number;
  createdAt: string;
}

export interface SubmitShadowDuelScoreBody {
  playerId: number;
  score: number;
  level: number;
  combo: number;
}

export type SyncUpSessionStatus =
  (typeof SyncUpSessionStatus)[keyof typeof SyncUpSessionStatus];

export const SyncUpSessionStatus = {
  waiting: "waiting",
  active: "active",
  completed: "completed",
} as const;

export interface SyncUpSession {
  id: number;
  player1Id: number;
  /** @nullable */
  player2Id?: number | null;
  status: SyncUpSessionStatus;
  /** @nullable */
  syncScore?: number | null;
  pattern: string;
  /** @nullable */
  player1Score?: number | null;
  /** @nullable */
  player2Score?: number | null;
  /** @nullable */
  pointsAwarded?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface CreateSyncUpSessionBody {
  player1Id: number;
  pattern: string;
}

export interface CompleteSyncUpBody {
  player2Id: number;
  player1Score: number;
  player2Score: number;
  syncScore: number;
}