# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Infinite Duo - Couples Game Hub API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all players
 */
export const GetPlayersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  avatar: zod.string(),
  totalPoints: zod.number(),
  weeklyPoints: zod.number(),
  gamesPlayed: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetPlayersResponse = zod.array(GetPlayersResponseItem);

/**
 * @summary Create or update a player
 */
export const CreatePlayerBody = zod.object({
  name: zod.string(),
  avatar: zod.string(),
});

/**
 * @summary Get player by id
 */
export const GetPlayerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPlayerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  avatar: zod.string(),
  totalPoints: zod.number(),
  weeklyPoints: zod.number(),
  gamesPlayed: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Get the weekly leaderboard
 */
export const GetLeaderboardResponse = zod.object({
  weeklyLeader: zod
    .object({
      playerId: zod.number(),
      playerName: zod.string(),
      avatar: zod.string(),
      weeklyPoints: zod.number(),
      totalPoints: zod.number(),
      rank: zod.number(),
    })
    .optional(),
  entries: zod.array(
    zod.object({
      playerId: zod.number(),
      playerName: zod.string(),
      avatar: zod.string(),
      weeklyPoints: zod.number(),
      totalPoints: zod.number(),
      rank: zod.number(),
    }),
  ),
  weekEndsAt: zod.coerce.date(),
});

/**
 * @summary Get the couple's shared relationship level and XP
 */
export const GetRelationshipLevelResponse = zod.object({
  level: zod.number(),
  levelName: zod.string(),
  currentXp: zod.number(),
  xpToNextLevel: zod.number(),
  totalXp: zod.number(),
  progressPercent: zod.number(),
});

/**
 * @summary Get aggregated game stats for the hub overview
 */
export const GetGameStatsResponse = zod.object({
  totalGamesPlayed: zod.number(),
  predictionsWon: zod.number(),
  predictionsTotal: zod.number(),
  shadowDuelTopScore: zod.number(),
  syncUpSessionsPlayed: zod.number(),
  currentStreak: zod.number(),
});

/**
 * @summary Get today's predictions
 */
export const GetPredictionsResponseItem = zod.object({
  id: zod.number(),
  playerId: zod.number(),
  playerName: zod.string(),
  question: zod.string(),
  prediction: zod.string(),
  betPoints: zod.number(),
  status: zod.enum(["pending", "correct", "wrong"]),
  resolvedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetPredictionsResponse = zod.array(GetPredictionsResponseItem);

/**
 * @summary Create a new daily prediction
 */
export const CreatePredictionBody = zod.object({
  playerId: zod.number(),
  question: zod.string(),
  prediction: zod.string(),
  betPoints: zod.number(),
});

/**
 * @summary Resolve a prediction (mark correct or wrong)
 */
export const ResolvePredictionParams = zod.object({
  id: zod.coerce.number(),
});

export const ResolvePredictionBody = zod.object({
  correct: zod.boolean(),
});

export const ResolvePredictionResponse = zod.object({
  id: zod.number(),
  playerId: zod.number(),
  playerName: zod.string(),
  question: zod.string(),
  prediction: zod.string(),
  betPoints: zod.number(),
  status: zod.enum(["pending", "correct", "wrong"]),
  resolvedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get prediction history with win rates
 */
export const GetPredictionHistoryResponse = zod.object({
  totalPredictions: zod.number(),
  correctPredictions: zod.number(),
  winRate: zod.number(),
  recentPredictions: zod.array(
    zod.object({
      id: zod.number(),
      playerId: zod.number(),
      playerName: zod.string(),
      question: zod.string(),
      prediction: zod.string(),
      betPoints: zod.number(),
      status: zod.enum(["pending", "correct", "wrong"]),
      resolvedAt: zod.coerce.date().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Get all shadow duel high scores
 */
export const GetShadowDuelScoresResponseItem = zod.object({
  id: zod.number(),
  playerId: zod.number(),
  playerName: zod.string(),
  avatar: zod.string(),
  score: zod.number(),
  level: zod.number(),
  combo: zod.number(),
  createdAt: zod.coerce.date(),
});
export const GetShadowDuelScoresResponse = zod.array(
  GetShadowDuelScoresResponseItem,
);

/**
 * @summary Submit a new shadow duel score
 */
export const SubmitShadowDuelScoreBody = zod.object({
  playerId: zod.number(),
  score: zod.number(),
  level: zod.number(),
  combo: zod.number(),
});

/**
 * @summary Get top 10 shadow duel scores
 */
export const GetShadowDuelLeaderboardResponseItem = zod.object({
  id: zod.number(),
  playerId: zod.number(),
  playerName: zod.string(),
  avatar: zod.string(),
  score: zod.number(),
  level: zod.number(),
  combo: zod.number(),
  createdAt: zod.coerce.date(),
});
export const GetShadowDuelLeaderboardResponse = zod.array(
  GetShadowDuelLeaderboardResponseItem,
);

/**
 * @summary Get sync-up game sessions
 */
export const GetSyncUpSessionsResponseItem = zod.object({
  id: zod.number(),
  player1Id: zod.number(),
  player2Id: zod.number().nullish(),
  status: zod.enum(["waiting", "active", "completed"]),
  syncScore: zod.number().nullish(),
  pattern: zod.string(),
  player1Score: zod.number().nullish(),
  player2Score: zod.number().nullish(),
  pointsAwarded: zod.number().nullish(),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullish(),
});
export const GetSyncUpSessionsResponse = zod.array(
  GetSyncUpSessionsResponseItem,
);

/**
 * @summary Create a new sync-up session
 */
export const CreateSyncUpSessionBody = zod.object({
  player1Id: zod.number(),
  pattern: zod.string(),
});

/**
 * @summary Complete a sync-up session with results
 */
export const CompleteSyncUpSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const CompleteSyncUpSessionBody = zod.object({
  player2Id: zod.number(),
  player1Score: zod.number(),
  player2Score: zod.number(),
  syncScore: zod.number(),
});

export const CompleteSyncUpSessionResponse = zod.object({
  id: zod.number(),
  player1Id: zod.number(),
  player2Id: zod.number().nullish(),
  status: zod.enum(["waiting", "active", "completed"]),
  syncScore: zod.number().nullish(),
  pattern: zod.string(),
  player1Score: zod.number().nullish(),
  player2Score: zod.number().nullish(),
  pointsAwarded: zod.number().nullish(),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullish(),
});